Unwanted extension check was overly aggressively deleting folders

+++ sabnzbd/filesystem.py
@@ -635,10 +635,13 @@ def cleanup_empty_directories(path):
                     pass
         if not repeat:
             break
-    try:
-        remove_dir(path)
-    except:
-        pass
+
+    # Only remove if main folder is now also empty
+    if not os.listdir(path):
+        try:
+            remove_dir(path)
+        except:
+            pass
 
 
 @synchronized(DIR_LOCK)
